add image to api

+++ Hotel/src/main/java/com/example/Hotel/repositories/ChambreRepository.java
@@ -3,5 +3,5 @@ package com.example.Hotel.repositories;
 import com.example.Hotel.model.hotel.Chambre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChambreRepository extends JpaRepository<Chambre, Long> {
+public interface ChambreRepository extends JpaRepository<Chambre, Integer> {
 }
